+++ src/components/HeaderTextContainer/HeaderTextContainer.stories.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import { HeaderTextContainer } from './HeaderTextContainer';
-
-const stories = storiesOf('HeaderTextContainer', module);
-stories.add('HeaderTextContainer', () => <HeaderTextContainer />);