+++ src/components/IncidentTextPanel/index.tsx
@@ -1,1 +1,0 @@
-export { IncidentTextPanel } from './IncidentTextPanel';