+++ src/components/Incidents/index.ts
@@ -1,0 +1,7 @@
+export { EmptyResults } from './EmptyResults';
+export { Incident } from './Incident';
+export { IncidentImage } from './IncidentImage';
+export { IncidentTextBlock } from './IncidentTextBlock';
+export { Pagination } from './Pagination';
+export { SearchPanel } from './SearchPanel';
+export { TotalIncidentsPanel } from './TotalIncidentsPanel';