+++ src/components/common/HeaderTextBlock/HeaderTextBlock.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { HeaderTextBlock } from './HeaderTextBlock';
+
+const stories = storiesOf('HeaderTextBlock', module);
+stories.add('HeaderTextBlock', () => <HeaderTextBlock />);