+++ src/components/common/HeaderTextBlock/index.tsx
@@ -1,0 +1,1 @@
+export { HeaderTextBlock } from './HeaderTextBlock';