+++ src/components/common/index.ts
@@ -1,0 +1,5 @@
+export { Error } from './Error';
+export { Header } from './Header';
+export { HeaderTextBlock } from './HeaderTextBlock';
+export { Layout } from './Layout';
+export { Loading } from './Loading';