+++ src/components/index.ts
@@ -1,14 +1,3 @@
-export { EmptyResults } from './EmptyResults';
-export { Error } from './Error';
-export { Header } from './Header';
-export { HeaderTextContainer } from './HeaderTextContainer';
-export { Incident } from './Incident';
-export { IncidentImage } from './IncidentImage';
-export { Incidents } from './Incidents';
-export { IncidentTextPanel } from './IncidentTextPanel';
-export { Layout } from './Layout';
-export { Loading } from './Loading';
-export { Map } from './Map';
-export { Pagination } from './Pagination';
-export { Search } from './Search';
-export { TotalIncidents } from './TotalIncidents';
+export * from './common';
+export * from './incidents';
+export * from './incidentDetail';